storing transactions to Firestone database and structuring the program
